MJCalculator
Prints out a list of the best locations for the player to be in and the best player-mj combos statistically based on average distance between the player and the switch. Avoids any paths intersecting a square MJ is on. If Player and MJ are on the same square, it auto-sets as "6" as a way to deter